refine, mv popVerfier to core/vm, rm pop in test genesis, fatal if consensus key is missing.
