+++ cmd/autonity/genesis_test.go
@@ -56,7 +56,6 @@ var genesisTest = struct {
 							"enode": "enode://395f3b74b236bccde1684d50a715c2349ee66741d7a69a571c0b176d129ac4ee33acbc85456f7ada3cdc87e1a56f591ba624870a1381c266d41b34c9476c5bc4@172.25.0.11:30303",
 							"treasury": "0x4b07239bd581d21aefcdee0c6db38070f9a5fd2d",
 							"consensusKey": "0xa5b7ba3a6abd4e6d2859733447e39996b8bb4d284aefc8e0f7a909d51873c13fa5c60246c96705a5ae239153f366f188",
-							"pop": "0x4cb3af04550c08dbb7dd0553fd735e3aed6d122dd9a2bf527bbf3af2354add5743f0ff6a6b6f27d03fb01a0cc3cdd390ea38fdb40192f0ef1a8386d0538bb2df01ac0555893eaedf6f605c178878492c7d95f26d706c0b84c5bb96b77f7e8428ef6ad143121317d5f48101d476d818911636eb0f5c25758932f8feddd61682f8d50096a0ffdc430d230f29db673f5e0eb6c726a8525bd2e997a92a2466b0d4f7d346874f22299edb072659de8b0f6186086113f221fa0d8ec30040db87dbbe9b5276c82dd2f11a2f9d286896e1181adb6f4744b9258c6ba5935ef1f8148271c5d313",
 							"oracleAddress": "0x342dbBAF7e5E41078Cb83905793CC28D3B7f5AD9",
 							"bondedStake" : 1
 						}
@@ -114,7 +113,6 @@ var genesisTest = struct {
 							"enode": "enode://395f3b74b236bccde1684d50a715c2349ee66741d7a69a571c0b176d129ac4ee33acbc85456f7ada3cdc87e1a56f591ba624870a1381c266d41b34c9476c5bc4@172.25.0.11:30303",
 							"treasury": "0x4b07239bd581d21aefcdee0c6db38070f9a5fd2d",
 							"consensusKey": "0xa5b7ba3a6abd4e6d2859733447e39996b8bb4d284aefc8e0f7a909d51873c13fa5c60246c96705a5ae239153f366f188",
-							"pop": "0x4cb3af04550c08dbb7dd0553fd735e3aed6d122dd9a2bf527bbf3af2354add5743f0ff6a6b6f27d03fb01a0cc3cdd390ea38fdb40192f0ef1a8386d0538bb2df01ac0555893eaedf6f605c178878492c7d95f26d706c0b84c5bb96b77f7e8428ef6ad143121317d5f48101d476d818911636eb0f5c25758932f8feddd61682f8d50096a0ffdc430d230f29db673f5e0eb6c726a8525bd2e997a92a2466b0d4f7d346874f22299edb072659de8b0f6186086113f221fa0d8ec30040db87dbbe9b5276c82dd2f11a2f9d286896e1181adb6f4744b9258c6ba5935ef1f8148271c5d313",
 							"oracleAddress": "0x342dbBAF7e5E41078Cb83905793CC28D3B7f5AD9",
 							"bondedStake" : 2
 						}
@@ -150,7 +148,6 @@ var genesisTest = struct {
 							"enode": "enode://395f3b74b236bccde1684d50a715c2349ee66741d7a69a571c0b176d129ac4ee33acbc85456f7ada3cdc87e1a56f591ba624870a1381c266d41b34c9476c5bc4@172.25.0.11:30303",
 							"treasury": "0x4b07239bd581d21aefcdee0c6db38070f9a5fd2d",
 							"consensusKey": "0xa5b7ba3a6abd4e6d2859733447e39996b8bb4d284aefc8e0f7a909d51873c13fa5c60246c96705a5ae239153f366f188",
-							"pop": "0x4cb3af04550c08dbb7dd0553fd735e3aed6d122dd9a2bf527bbf3af2354add5743f0ff6a6b6f27d03fb01a0cc3cdd390ea38fdb40192f0ef1a8386d0538bb2df01ac0555893eaedf6f605c178878492c7d95f26d706c0b84c5bb96b77f7e8428ef6ad143121317d5f48101d476d818911636eb0f5c25758932f8feddd61682f8d50096a0ffdc430d230f29db673f5e0eb6c726a8525bd2e997a92a2466b0d4f7d346874f22299edb072659de8b0f6186086113f221fa0d8ec30040db87dbbe9b5276c82dd2f11a2f9d286896e1181adb6f4744b9258c6ba5935ef1f8148271c5d313",
 							"oracleAddress": "0x342dbBAF7e5E41078Cb83905793CC28D3B7f5AD9",
 							"bondedStake" : 1
 						}
