+++ core/vm/contracts_test.go
@@ -20,6 +20,9 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
+	"github.com/autonity/autonity/crypto"
+	"github.com/autonity/autonity/crypto/blst"
+	"github.com/stretchr/testify/require"
 	"io/ioutil"
 	"testing"
 	"time"
@@ -392,4 +395,35 @@ func BenchmarkPrecompiledBLS12381G2MultiExpWorstCase(b *testing.B) {
 		NoBenchmark: false,
 	}
 	benchmarkPrecompiled("0f", testcase, b)
+}
+
+func TestPOPVerifier(t *testing.T) {
+	key1, err := crypto.GenerateKey()
+	require.NoError(t, err)
+
+	treasuryAddress := crypto.PubkeyToAddress(key1.PublicKey)
+	key, err := blst.RandKey()
+	require.NoError(t, err)
+
+	proof, err := crypto.BLSPOPProof(key, treasuryAddress.Bytes())
+	require.NoError(t, err)
+
+	popVerifier := &POPVerifier{}
+	input := make([]byte, ArrayLenBytes+POPBytes)
+	signatureOffset := ArrayLenBytes + blst.BLSPubkeyLength
+	treasuryOffset := signatureOffset + blst.BLSSignatureLength
+	copy(input[ArrayLenBytes:signatureOffset], key.PublicKey().Marshal())
+	copy(input[signatureOffset:treasuryOffset], proof)
+	copy(input[treasuryOffset:ArrayLenBytes+POPBytes], treasuryAddress.Bytes())
+
+	ret, err := popVerifier.Run(input, 0)
+	require.NoError(t, err)
+	require.Equal(t, successResult, ret)
+
+	wrongKey, err := blst.RandKey()
+	require.NoError(t, err)
+	copy(input[32:80], wrongKey.PublicKey().Marshal())
+	ret, err = popVerifier.Run(input, 0)
+	require.NotNil(t, err)
+	require.Equal(t, failure32Byte, ret)
 }
